+++ config/routes.rb
@@ -8,16 +8,20 @@ Rails.application.routes.draw do
       delete '/logout', to: 'sessions#destroy'
       get '/logged_in', to: 'sessions#logged_in?'
       # ユーザー登録、プロフィール更新、パスワード更新、フォロー、フォロー解除
-      resources :users, only: %i[index create], module: :users do
-        resource :profile, only: %i[update], controller: :profile
-        resource :password, only: %i[update], controller: :password
-        resources :relationships, only: %i[create destroy]
+      resources :users, only: %i[index create] do
+        scope module: :users do
+          resource :profile, only: %i[update], controller: :profile
+          resource :password, only: %i[update], controller: :password
+          resources :relationships, only: %i[create destroy]
+        end
       end
 
       resources :posts, only: %i[index create destroy]
 
       resources :images, only: %i[index destroy] do
-        resources :likes, only: %i[create destroy]
+        scope module: :images do
+          resources :likes, only: %i[create destroy]
+        end
       end
       put '/images/bulk_update', to: 'images#bulk_update' # 一括更新
